LayoutTests/media/video-muted.html

<video controls></video>
<p>Test 'muted' attribute<p>
<script src=video-test.js></script>
<script>
test("!video.muted");
run("video.muted = true");
test("video.muted");
video.src = "content/test.mp4";
run("video.load()");
waitForEvent("load", function () {
    test("video.muted");
    run("video.muted = false");
    test("!video.muted");
    endTest();
});
</script>
